requirements.txt
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
DateTime==4.3
et-xmlfile==1.1.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.2
openpyxl==3.0.7
pandas==1.3.2
Pillow==8.3.1
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
zope.interface==5.4.0
